EA3152: Rename files for better conformance to naming conventions
