+++ nuttx/configs/ea3152/src/lpc31_buttons.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/ea3152/src/up_leds.c
+ * configs/ea3152/src/lpc31_buttons.c
  *
  *   Copyright (C) 2011, 2014-2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
