+++ nuttx/configs/ea3152/src/lpc31_mem.c
@@ -1,6 +1,5 @@
 /****************************************************************************
- * configs/ea3152/src/up_mem.c
- * arch/arm/src/board/up_mem.c
+ * configs/ea3152/src/lpc31_mem.c
  *
  *   Copyright (C) 2011 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
